fix! test folders refactor
