+++ web-api/tests/Kerbero.Integration.Tests/SmartLocks/NukiSmartLocksListIntegrationTests.cs
@@ -4,16 +4,9 @@ using System.Text.Json;
 using System.Text.Json.Nodes;
 using FluentAssertions;
 using Flurl.Http.Testing;
-using Kerbero.Infrastructure.Common.Context;
 using Kerbero.WebApi;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
 
-namespace Kerbero.Integration.Tests;
+namespace Kerbero.Integration.Tests.SmartLocks;
 
 public class NukiSmartLocksListIntegrationTests: IDisposable
 {
