+++ web-api/tests/Kerbero.WebApi.Tests/Utils/ModelStateDictionaryExtensionsTest.cs
@@ -1,13 +1,10 @@
 using System.Net;
 using FluentAssertions;
-using FluentResults;
 using Kerbero.Domain.Common.Errors;
-using Kerbero.Domain.NukiAuthentication.Models;
 using Kerbero.WebApi.Utils.Extensions;
 using Microsoft.AspNetCore.Mvc;
-using Moq;
 
-namespace Kerbero.WebApi.Tests;
+namespace Kerbero.WebApi.Tests.Utils;
 
 public class ModelStateDictionaryExtensionsTest: ControllerBase
 {
